Adding Azure VM read permission to PKS workers
This permission seems to be required for provisioning PersistentVolumes
when using a StorageClass that uses Azure managed disks. This feature
only became available in Azure v1.7.2, so it's possible our old testing
only included the old StorageClass configuration.

If this permission is missing from the k8s workers, Pods will not be able
to mount PersistentVolumes, and an error will appear in the k8s events
saying that the worker VM "does not have authorization to perform action
'Microsoft.Compute/virtualMachines/read'" on the disk.